backend/requirements.txt
supabase
google-generativeai
feedparser
requests
python-dotenv